backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pymavlink
pandas
pyarrow
redis
python-multipart
pydantic